+++ lib/vault-tree/keywords/public_encryption_key.rb
@@ -15,8 +15,8 @@ module VaultTree
     def decryption_key
       begin
         contract.retrieve_contents(vault_id)
-      rescue Exceptions::EmptyVault
-        raise Exceptions::MissingPartnerDecryptionKey
+      rescue Exceptions::EmptyVault => e
+        raise Exceptions::MissingPartnerDecryptionKey.new(e, vault_id: vault_id )
       end
     end
   end